ShashChess 30 Solved a bug in winProbability implementation (replaced the map with an array). Improvements in Shashin theory. New net from stockfish testing framework, slightly better than the original one. Stockfish patch: Remove 'si' StateInfo variable/parameter. Since st is a member of position we don't need to pass it separately as parameter.
While being there also remove some line in pos_is_ok, where
a copy of StateInfo was made by using default copy constructor and
then verified it's correctedness by doing a memcmp.
There is no point in doing that.

+++ src/All/uci.h
@@ -81,7 +81,7 @@ std::string square(Square s);
 std::string move(Move m, bool chess960);
 std::string pv(const Position& pos, Depth depth);
 std::string wdl(Value v, int ply);
-int getWinProbability(Value v, int ply);
+uint8_t getWinProbability(Value v, int ply);
 Move to_move(const Position& pos, std::string& str);
 
 } // namespace UCI
